refactor(typo): fix typo in PAM types file name
