+++ lib/core/types/api/access-manager.js
