+++ lib/types/core/types/api/access-manager.d.ts
