+++ src/core/endpoints/access_manager/audit.ts
@@ -7,7 +7,7 @@ import { TransportResponse } from '../../types/transport-response';
 import { PubNubAPIError } from '../../../errors/pubnub-api-error';
 import { AbstractRequest } from '../../components/request';
 import RequestOperation from '../../constants/operations';
-import * as PAM from '../../types/api/access-panager';
+import * as PAM from '../../types/api/access-manager';
 import { KeySet, Query } from '../../types/api';
 
 // --------------------------------------------------------
