+++ src/core/types/api/access-manager.ts
